+++ src/main/java/ru/yandex/practicum/filmorate/service/film/FilmService.java
@@ -36,7 +36,7 @@ public interface FilmService {
 
     void deleteFilm(long id);
 
-    List<Film> getFilmsByDirector(Long directorId, String sortBy);
+    List<Film> getFilmsByDirector(Integer directorId, String sortBy);
 
     List<Film> getCommonFilms(@Positive(message = "id должен быть положительным") Long userId,
                               @Positive(message = "id должен быть положительным") Long friendId);
