+++ src/main/java/ru/yandex/practicum/filmorate/storage/FilmStorage.java
@@ -33,7 +33,7 @@ public interface FilmStorage {
 
     List<Film> returnFilmsListByIDs(List<Long> filmsIds);
 
-    List<Film> getFilmsByDirector(Long directorId, String sortBy);
+    List<Film> getFilmsByDirector(Integer directorId, String sortBy);
 
     List<Film> getFilmsAfterSearching(String query, String by);
 }
