+++ src/main/java/ru/yandex/practicum/filmorate/storage/memory/film/InMemoryFilmStorage.java
@@ -131,7 +131,7 @@ public class InMemoryFilmStorage implements FilmStorage {
     }
 
     @Override
-    public List<Film> getFilmsByDirector(Long directorId, String sortBy) {
+    public List<Film> getFilmsByDirector(Integer directorId, String sortBy) {
         return new ArrayList<>();
     }
